+++ packages/scene-graph/src/node-defaults.ts
@@ -30,7 +30,8 @@ export function createDefaultNode(
         componentPropAssignments: [],
         derivedSymbolData: [],
         derivedSymbolDataLayoutVersion: null,
-        uniformScaleFactor: null
+        uniformScaleFactor: null,
+        kiwiNodeType: null
       }
     },
     figmaDerivedLayout: null,